+++ Graph.h
@@ -1,3 +1,9 @@
+/**
+* @file Graph.h
+ * @brief Implementation of a Graph data structure representation, based on the one
+ *        provided in the classes.
+ */
+
 #ifndef GRAPH_H
 #define GRAPH_H
 
@@ -7,6 +13,14 @@
 
 class Edge;
 
+/**
+ * @brief Class representing a vertex in the graph.
+ *
+ * A vertex represents a node in the graph that can have outgoing and incoming edges,
+ * and can store various attributes such as location, ID, parking availability, and others.
+ * Vertices are used in graph traversal algorithms, and they store information about their
+ * adjacent vertices (edges).
+ */
 class Vertex {
 protected:
     std::string location;
@@ -18,7 +32,7 @@ protected:
     bool processing;
     int low = -1, num = -1;
     unsigned int indegree;
-    double dist;
+    long dist;
     Edge * path = nullptr;
 
     std::vector<Edge *> adj;
@@ -42,8 +56,8 @@ public:
     void setProcessing(bool processing);
     unsigned int getIndegree() const;
     void setIndegree(unsigned int indegree);
-    double getDist() const;
-    void setDist(double dist);
+    long getDist() const;
+    void setDist(long dist);
     int getLow() const;
     void setLow(int value);
     int getNum() const;
@@ -63,7 +77,12 @@ public:
     int queueIndex = 0;
 };
 
-
+/**
+ * @brief Class representing an edge in the graph.
+ *
+ * An edge connects two vertices and can store information about the edge such as distance,
+ * flow, and whether it's selected. An edge also has a reverse edge to handle bidirectional connections.
+ */
 class Edge {
 protected:
     Vertex *orig;
@@ -91,7 +110,12 @@ public:
 };
 
 
-
+/**
+ * @brief Class representing a graph with vertices and edges.
+ *
+ * A graph consists of a set of vertices and edges connecting them. The class provides methods
+ * for adding vertices and edges, as well as finding vertices by ID or code.
+ */
 class Graph {
 protected:
     std::vector<Vertex *> vertexSet;
